+++ packages/graphql/src/graphql-bridge.ts
@@ -3,5 +3,4 @@ import { builder } from "./pothos";
 
 export const graphqlBridge = initializeDrizzleGraphqlBridge(builder, {
   defaultIdFields: ["id", "orgId"],
-  unmappedFields: "throw",
 });